refactor(web,mobile,worker): improve separation of concerns across packages
- Extract useFileUpload hook from ScoresheetPanel (447→280 LOC component)
- Extract useValidationActions from useValidationState (457→280 LOC hook)
- Migrate mobile departureReminders store from custom pub/sub to Zustand
- Deduplicate API_BASE_URL constant in mobile (2 files → 1 shared constant)
- Replace export * barrel files with explicit exports in validation/settings
- Harmonize ESLint magic numbers rule for worker package
- Extract iOS Safari session relay utilities from worker proxy handler
