+++ web-app/knip.json
@@ -2,7 +2,7 @@
   "$schema": "https://unpkg.com/knip@5/schema.json",
   "entry": ["src/main.tsx", "e2e/**/*.ts", "src/features/*/index.ts"],
   "project": ["src/**/*.{ts,tsx}", "e2e/**/*.ts"],
-  "ignore": ["src/types/**/*.d.ts", "src/api/schema.ts"],
+  "ignore": ["src/types/**/*.d.ts", "src/api/schema.ts", "src/features/validation/components/OCRPanel.tsx"],
   "ignoreDependencies": [
     "tailwindcss",
     "@volleykit/shared",
