+++ web-app/src/features/settings/index.ts
@@ -9,8 +9,5 @@ export {
   useSuspenseActiveSeason,
 } from './hooks/useSettings'
 
-// Components
-export * from './components'
-
 // Tour
 export { settingsTour } from './settings'
