+++ web-app/src/features/validation/hooks/useValidationState.ts
@@ -1,15 +1,13 @@
 import { useState, useCallback, useMemo, useRef, useEffect } from 'react'
 
-import { useQuery, useQueryClient } from '@tanstack/react-query'
+import { useQuery } from '@tanstack/react-query'
 
 import { getApiClient } from '@/api/client'
-import { queryKeys } from '@/api/queryKeys'
 import { gameDetailOptions } from '@/api/queryOptions'
 import type { ValidatedPersonSearchResult } from '@/api/validation'
 import type { RosterPanelModifications } from '@/features/validation/components/RosterVerificationPanel'
 import { useAuthStore } from '@/shared/stores/auth'
 import type { PendingScorerData } from '@/shared/stores/demo'
-import { logger } from '@/shared/utils/logger'
 
 import {
   type ValidationState,
@@ -18,13 +16,8 @@ import {
   type UseValidationStateResult,
   createInitialState,
 } from './types'
-import {
-  hasRosterModifications,
-  saveRosterModifications,
-  saveScorerSelection,
-  finalizeRoster,
-  finalizeScoresheetWithFile,
-} from '../api/api-helpers'
+import { useValidationActions } from './useValidationActions'
+import { hasRosterModifications } from '../api/api-helpers'
 
 // Re-export types for consumers
 export type {
@@ -57,15 +50,9 @@ export type { PendingScorerData as PendingScorerInfo } from '@/shared/stores/dem
  */
 export function useValidationState(gameId?: string): UseValidationStateResult {
   const [state, setState] = useState<ValidationState>(createInitialState)
-  const [isSaving, setIsSaving] = useState(false)
-  const [isFinalizing, setIsFinalizing] = useState(false)
-  const isSavingRef = useRef(false)
-  const isFinalizingRef = useRef(false)
-  const uploadedFileResourceIdRef = useRef<string | undefined>(undefined)
 
   const dataSource = useAuthStore((s) => s.dataSource)
   const apiClient = getApiClient(dataSource)
-  const queryClient = useQueryClient()
 
   // Fetch game details (scoresheet and nomination list IDs)
   const gameDetailsQuery = useQuery({
@@ -166,16 +153,6 @@ export function useValidationState(gameId?: string): UseValidationStateResult {
     [gameDetailsQuery.data?.scoresheet?.file]
   )
 
-  // Pre-fill the uploaded file resource ID ref when an existing scoresheet file exists.
-  // This ensures that save/finalize operations reuse the existing file reference
-  // instead of requiring a new upload when the user hasn't changed the scoresheet.
-  // Skip when scoresheet is not required — file references should never be sent for those games.
-  useEffect(() => {
-    if (existingFileResourceId && !uploadedFileResourceIdRef.current && !scoresheetNotRequired) {
-      uploadedFileResourceIdRef.current = existingFileResourceId
-    }
-  }, [existingFileResourceId, scoresheetNotRequired])
-
   const isDirty = useMemo(() => {
     return (
       hasRosterModifications(state.homeRoster.playerModifications) ||
@@ -272,158 +249,14 @@ export function useValidationState(gameId?: string): UseValidationStateResult {
     }
   }, [])
 
-  const saveProgress = useCallback(async (): Promise<void> => {
-    if (isSavingRef.current) {
-      logger.debug('[VS] save skip: in progress')
-      return
-    }
-
-    isSavingRef.current = true
-    setIsSaving(true)
-
-    try {
-      const gameDetails = gameDetailsQuery.data
-      if (!gameId || !gameDetails) {
-        logger.warn('[VS] no game data for save')
-        return
-      }
-
-      // DIAGNOSTIC: log scoresheet state to debug missing PUT/POST (see #924)
-      logger.debug('[VS] saveProgress: scoresheet data from API', {
-        hasScoresheet: !!gameDetails.scoresheet,
-        scoresheetId: gameDetails.scoresheet?.__identity,
-        scorerId: state.scorer.selected?.__identity,
-        hasNoScoresheet: gameDetails.group?.hasNoScoresheet,
-      })
-
-      // Upload scoresheet file if present (cache the resource ID to avoid re-upload on finalize)
-      // Skip upload and file reference entirely when scoresheet is not required for this group
-      let fileResourceId: string | undefined
-      if (!scoresheetNotRequired) {
-        fileResourceId = uploadedFileResourceIdRef.current
-        if (!fileResourceId && state.scoresheet.file) {
-          const uploadResult = await apiClient.uploadResource(state.scoresheet.file)
-          fileResourceId = uploadResult[0]?.__identity
-          uploadedFileResourceIdRef.current = fileResourceId
-          logger.debug('[VS] PDF uploaded:', fileResourceId)
-        }
-      }
-
-      await saveRosterModifications(
-        apiClient,
-        gameId,
-        gameDetails.nominationListOfTeamHome,
-        state.homeRoster.playerModifications,
-        state.homeRoster.coachModifications
-      )
-      await saveRosterModifications(
-        apiClient,
-        gameId,
-        gameDetails.nominationListOfTeamAway,
-        state.awayRoster.playerModifications,
-        state.awayRoster.coachModifications
-      )
-      await saveScorerSelection(
-        apiClient,
-        gameId,
-        gameDetails.scoresheet,
-        state.scorer.selected?.__identity,
-        fileResourceId
-      )
-
-      // Invalidate cache so reopening shows the saved data
-      await queryClient.invalidateQueries({ queryKey: queryKeys.validation.gameDetail(gameId) })
-      logger.debug('[VS] save done')
-    } catch (error) {
-      logger.error('[VS] save failed:', error)
-      throw error
-    } finally {
-      isSavingRef.current = false
-      setIsSaving(false)
-    }
-  }, [gameId, gameDetailsQuery.data, state, apiClient, queryClient, scoresheetNotRequired])
-
-  const finalizeValidation = useCallback(async (): Promise<void> => {
-    if (isFinalizingRef.current) {
-      logger.debug('[VS] finalize skip: in progress')
-      return
-    }
-
-    isFinalizingRef.current = true
-    setIsFinalizing(true)
-
-    try {
-      const gameDetails = gameDetailsQuery.data
-      if (!gameId || !gameDetails) {
-        logger.warn('[VS] no game data for finalize')
-        return
-      }
-
-      // Reuse cached file resource ID from saveProgress if available
-      // Skip upload and file reference entirely when scoresheet is not required for this group
-      let fileResourceId: string | undefined
-      if (!scoresheetNotRequired) {
-        fileResourceId = uploadedFileResourceIdRef.current
-        if (!fileResourceId && state.scoresheet.file) {
-          const uploadResult = await apiClient.uploadResource(state.scoresheet.file)
-          fileResourceId = uploadResult[0]?.__identity
-          uploadedFileResourceIdRef.current = fileResourceId
-          logger.debug('[VS] PDF uploaded:', fileResourceId)
-        }
-      }
-
-      await finalizeRoster(
-        apiClient,
-        gameId,
-        gameDetails.nominationListOfTeamHome,
-        state.homeRoster.playerModifications,
-        state.homeRoster.coachModifications
-      )
-      await finalizeRoster(
-        apiClient,
-        gameId,
-        gameDetails.nominationListOfTeamAway,
-        state.awayRoster.playerModifications,
-        state.awayRoster.coachModifications
-      )
-
-      // Save scorer/scoresheet first to ensure the scoresheet record exists on the
-      // server. Without safe-mode, saveProgress() is never called so the scoresheet
-      // may not have been created yet, causing finalizeScoresheetWithFile to silently
-      // skip due to a missing scoresheet identity.
-      await saveScorerSelection(
-        apiClient,
-        gameId,
-        gameDetails.scoresheet,
-        state.scorer.selected?.__identity,
-        fileResourceId
-      )
-
-      // Re-fetch game details to pick up the scoresheet identity created by the save
-      const freshGameDetails = await apiClient.getGameWithScoresheet(gameId)
-      if (!freshGameDetails.scoresheet?.__identity) {
-        throw new Error('Scoresheet was not created after save — cannot finalize')
-      }
-
-      await finalizeScoresheetWithFile(
-        apiClient,
-        gameId,
-        freshGameDetails.scoresheet,
-        state.scorer.selected?.__identity,
-        fileResourceId
-      )
-
-      await queryClient.invalidateQueries({ queryKey: queryKeys.validation.gameDetail(gameId!) })
-      await queryClient.invalidateQueries({ queryKey: queryKeys.assignments.lists() })
-      logger.debug('[VS] finalize done')
-    } catch (error) {
-      logger.error('[VS] finalize failed:', error)
-      throw error
-    } finally {
-      isFinalizingRef.current = false
-      setIsFinalizing(false)
-    }
-  }, [gameId, gameDetailsQuery.data, state, apiClient, queryClient, scoresheetNotRequired])
+  // Delegate API operations to the actions hook
+  const { saveProgress, finalizeValidation, isSaving, isFinalizing } = useValidationActions({
+    gameId,
+    state,
+    scoresheetNotRequired,
+    gameDetails: gameDetailsQuery.data,
+    existingFileResourceId,
+  })
 
   return {
     state,
