+++ web-app/src/features/validation/index.ts
@@ -4,6 +4,3 @@ export { ValidateGameModal } from './components/ValidateGameModal'
 // Hooks
 export { useValidateGameWizard } from './hooks/useValidateGameWizard'
 export { useValidationState } from './hooks/useValidationState'
-
-// Components
-export * from './components'
