+++ worker/src/utils.ts
@@ -13,4 +13,5 @@ export * from './utils/cookie'
 export * from './utils/session'
 export * from './utils/cache'
 export * from './utils/ical-validation'
+export * from './utils/session-relay'
 export * from './utils/constants'
